runner/requirements.txt
fastapi
uvicorn
llama-cpp-python
huggingface-hub
